feature/ADAPT-3752 moved helper functions to a new file
